chore: refactoring for new notification badge component
